pyproject.toml
[tool.poetry]
name = "renderables"
version = "0.1.0"
description = "build your own (toy?) static site generator"
authors = ["Marcel Langer <dev@marcel.computer>"]

[tool.poetry.dependencies]
python = "^3.7"
Mako = ">=1.1.5"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
